backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
authlib==1.3.2
itsdangerous==2.2.0
httpx==0.27.2
starlette==0.38.6
google-genai==1.1.0
google-generativeai
python-dotenv
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
beautifulsoup4
lxml
qdrant-client
redis
sec-edgar-downloader
pytest
pytest-asyncio==1.0.1